Feat: split test by hash type
