--- Example/Tests/Tests.swift
@@ -1,11 +1,0 @@
-// https://github.com/Quick/Quick
-
-import XCTest
-import Hasher
-
-class TestTest: XCTest {
-    
-    func testTest() {
-        XCTAssert(true)
-    }
-}
